// src/types/enums.ts
export enum LockerStatus {
    OPEN = 'OPEN',
    CLOSED = 'CLOSED',
}

export enum RentSize {
    XS = 'XS',
    S = 'S',
    M = 'M',
    L = 'L',
    XL = 'XL',
}

export enum RentStatus {
    CREATED = 'CREATED',
    WAITING_DROPOFF = 'WAITING_DROPOFF',
    WAITING_PICKUP = 'WAITING_PICKUP',
    DELIVERED = 'DELIVERED',
}